requirements.txt
absl-py==1.0.0
attrs==21.4.0
cachetools==4.2.4
certifi==2021.10.8
charset-normalizer==2.0.9
cycler==0.11.0
flake8==4.0.1
fonttools==4.28.5
google-auth==2.3.3
google-auth-oauthlib==0.4.6
grpcio==1.43.0
idna==3.3
importlib-metadata==4.10.0
iniconfig==1.1.1
kiwisolver==1.3.2
Markdown==3.3.6
matplotlib==3.5.1
mccabe==0.6.1
numpy==1.21.5
oauthlib==3.1.1
packaging==21.3
pandas==1.3.5
Pillow==8.4.0
pluggy==1.0.0
protobuf==3.19.1
py==1.11.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.8.0
pyflakes==2.4.0
pyparsing==3.0.6
pytest==6.2.5
python-dateutil==2.8.2
pytz==2021.3
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.8
scipy==1.7.3
seaborn==0.11.2
six==1.16.0

tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
toml==0.10.2

torch-tb-profiler==0.3.1

tqdm==4.62.3
typing_extensions==4.0.1
urllib3==1.26.7
Werkzeug==2.0.2
zipp==3.6.0
